requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==1.17.1
cryptography==45.0.6
markdown2==2.5.4
pillow==11.3.0
pycparser==2.22
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
shiboken6==6.9.1

#python 3.12.4
